Moved the MLNExporter code into its own Maven build.
- Moved the MLNExporter code into its own Maven directory structure.
- Updated the aggregate POM file.
- In the future we might want to move the MLNExporter code into a
  separate repository.
